error control score
error control score float to int

+++ game2/lives.pde
@@ -1,4 +1,4 @@
-float score =0;
+int score =0;
 
 void lose(){
   println("noooo");
